+++ src/app/routes/routes.module.ts
@@ -6,11 +6,14 @@ import { LayoutComponent } from '../layout.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { DemoComponent } from './demo/demo.component';
 
+import { PassportComponent } from '../passport.component';
+import { LoginComponent } from './passport/login.component';
+
 @NgModule({
     imports: [
         SharedModule,
         RouterModule.forRoot([
-            { path: 'demo', component: DemoComponent },
+            { path: 'demo', component: DemoComponent, data: { title: 'DEMO' } },
             {
                 path: '',
                 component: LayoutComponent,
@@ -21,12 +24,20 @@ import { DemoComponent } from './demo/demo.component';
                     { path: 'acl', loadChildren: './acl/acl.module#DEMOACLModule' },
                     { path: 'editor', loadChildren: './editor/editor.module#EditorModule' }
                 ]
+            },
+            {
+                path: 'login',
+                component: PassportComponent,
+                children: [
+                    { path: '', component: LoginComponent, data: { title: 'Login' } }
+                ]
             }
         ], { useHash: true })
     ],
     declarations: [
         DashboardComponent,
-        DemoComponent
+        DemoComponent,
+        LoginComponent
     ],
     exports: [
         RouterModule