+++ src/core/abc/reuse-tab/reuse-tab.service.ts
@@ -8,12 +8,15 @@ import { ReuseTabCached, ReuseTabMatchMode, ReuseTabNotify } from './interface';
 @Injectable()
 export class ReuseTabService implements OnDestroy {
     private _max = 10;
+    private _debug = false;
     private _mode = ReuseTabMatchMode.Menu;
     private _excludes: RegExp[] = [ ];
     private _cachedChange: BehaviorSubject<ReuseTabNotify> = new BehaviorSubject<ReuseTabNotify>(null);
     private _cached: ReuseTabCached[] = [];
     private _titleCached: { [url: string]: string } = {};
+    private _hookCached: { [url: string]: boolean } = {};
     private removeBuffer: string;
+    private curUrl: string;
 
     // region: public
 
@@ -27,6 +30,10 @@ export class ReuseTabService implements OnDestroy {
     /** 设置匹配模式 */
     set mode(value: ReuseTabMatchMode) {
         this._mode = value;
+    }
+    /** 设置Debug模式 */
+    set debug(value: boolean) {
+        this._debug = value;
     }
     /** 排除规则，限 `mode=URL` */
     set excludes(values: RegExp[]) {
@@ -50,10 +57,17 @@ export class ReuseTabService implements OnDestroy {
         if (_handle && _handle.componentRef && _handle.componentRef.destroy) _handle.componentRef.destroy();
     }
 
+    private di(...args) {
+        if (!this._debug || !console) return ;
+        // tslint:disable-next-line:no-console
+        console.warn(...args);
+    }
+
     /**
      * 根据URL移除标签
      */
     remove(url: string) {
+        this.di('remove tag', url);
         this._cachedChange.next({ active: 'remove', url });
     }
 
@@ -81,6 +95,7 @@ export class ReuseTabService implements OnDestroy {
      * 清除所有缓存
      */
     clear() {
+        this.di('clear all catch');
         this.removeBuffer = null;
         this._cached.forEach(v => this.destroy(v._handle));
         this._cached = [];
@@ -107,7 +122,9 @@ export class ReuseTabService implements OnDestroy {
     /** 设置当前页标题 */
     set title(value: string) {
         if (!value) return;
-        this._titleCached[this.getUrl(this.injector.get(ActivatedRoute).snapshot)] = value;
+        if (!this.curUrl) this.curUrl = this.getUrl(this.injector.get(ActivatedRoute).snapshot);
+        this._titleCached[this.curUrl] = value;
+        this.di('update current tag title', value);
         this._cachedChange.next({ active: 'title', title: value });
     }
 
@@ -159,6 +176,16 @@ export class ReuseTabService implements OnDestroy {
         return menus.pop();
     }
 
+    private runHook(method: string, url: string, comp: any) {
+        if (this._hookCached[url]) return;
+        this._hookCached[url] = true;
+        setTimeout(() => {
+            if (comp.instance && comp.instance[method])
+                comp.instance[method]();
+            this._hookCached[url] = false;
+        }, 100);
+    }
+
     /** @private */
     getClosable(url: string, route?: ActivatedRouteSnapshot): boolean {
         if (route && route.data && typeof route.data.reuseClosable !== 'undefined')
@@ -197,13 +224,14 @@ export class ReuseTabService implements OnDestroy {
      */
     shouldDetach(route: ActivatedRouteSnapshot): boolean {
         if (!route.routeConfig || route.routeConfig.loadChildren || route.routeConfig.children) return false;
+        this.di('#shouldDetach', this.getUrl(route), this.can(route));
         return this.can(route);
     }
 
     /**
      * 存储
      */
-    store(_snapshot: ActivatedRouteSnapshot, _handle: {}) {
+    store(_snapshot: ActivatedRouteSnapshot, _handle: any) {
         if (!_snapshot.routeConfig || _snapshot.routeConfig.loadChildren || _snapshot.routeConfig.children) return ;
         if (this.count >= this._max) this._cached.shift();
         const url = this.getUrl(_snapshot);
@@ -224,6 +252,12 @@ export class ReuseTabService implements OnDestroy {
         }
         this._clearRemoveBuffer();
 
+        this.di('#store', url, idx === -1 ? '[new]' : '[override]');
+
+        if (_handle && _handle.componentRef) {
+            this.runHook('_onReuseDestroy', url, _handle.componentRef);
+        }
+
         this._cachedChange.next({ active: 'add', item });
     }
 
@@ -235,6 +269,7 @@ export class ReuseTabService implements OnDestroy {
         const url = this.getUrl(route);
         const data = this.get(url);
         const ret = !!(data && data._handle);
+        this.di('#shouldAttach', url, ret);
         return ret;
     }
 
@@ -245,14 +280,32 @@ export class ReuseTabService implements OnDestroy {
         if (!route.routeConfig || route.routeConfig.loadChildren || route.routeConfig.children) return null;
         const url = this.getUrl(route);
         const data = this.get(url);
-        return (data && data._handle) || null;
+        const ret = (data && data._handle) || null;
+        this.di('#retrieve', url, ret);
+        if (ret && ret.componentRef) {
+            this.runHook('_onReuseInit', url, ret.componentRef);
+        }
+        return ret;
     }
 
     /**
      * 决定是否应该进行复用路由处理
      */
     shouldReuseRoute(future: ActivatedRouteSnapshot, curr: ActivatedRouteSnapshot): boolean {
-        return future.routeConfig === curr.routeConfig;
+        let ret = future.routeConfig === curr.routeConfig;
+        let url = '';
+        if (ret) {
+            const path = ((future.routeConfig && future.routeConfig.path) || '') as string;
+            if (path.length > 0 && ~path.indexOf(':')) {
+                const futureUrl = this.getUrl(future);
+                const currUrl = this.getUrl(curr);
+                url = futureUrl;
+                ret = futureUrl === currUrl;
+            }
+        }
+        this.curUrl = ret ? '' : (url || this.getUrl(curr));
+        this.di('#shouldReuseRoute', future, curr, ret);
+        return ret;
     }
 
     ngOnDestroy(): void {